Improved Usability
Got rid of the first text box and replaced it with a radio button, added
a label to the lower boxes, set it up so that it prompts for user for
non-epsilon transitions and potentially state names. Also fixed
autonumbering for states so it's more JFLAP-y. Transition creator
function now takes in a character as a parameter as well.